+++ link/src/main/java/app/Promocion.java
@@ -1,7 +1,7 @@
 package app;
 
 public interface Promocion {
-	public float descuento(Orden orden);
+	public Double descuento(Orden orden);
 	public void utilizar();
 
 }
